test(range): migrate item
