+++ core/src/components/range/test/item/range.e2e.ts-snapshots/range-inset-list-md-rtl-Mobile-Safari-linux.png
