+++ core/src/components/range/test/item/range.e2e.ts-snapshots/range-item-color-ios-ltr-Mobile-Firefox-linux.png
