+++ core/src/components/range/test/item/range.e2e.ts-snapshots/range-item-color-md-ltr-Mobile-Chrome-linux.png
