+++ core/src/components/range/test/item/range.e2e.ts-snapshots/range-list-ios-rtl-Mobile-Chrome-linux.png
