+++ core/src/components/range/test/item/range.e2e.ts-snapshots/range-list-md-rtl-Mobile-Safari-linux.png
